Fix startup connection string
If you want to solve a git mystery, check out the history surrounding
this and PR #5, and try to figure out why the commits by Chad do not
show up in `git log src/backend-api/Startup.cs` unless you use the
`--follow` flag of `git log`.

I wonder if there was a bad merge somewhere and have spent a bit of time
trying to understand the problem, but I have to move on. Hopefully this
is the only bad thing that happened with whatever actually happened.

Good luck. We're all counting on you.